carrot_impl: refactor fake_pruned_blockchain into header
