+++ tests/unit_tests/CMakeLists.txt
@@ -60,6 +60,7 @@ set(unit_tests_sources
   epee_serialization.cpp
   epee_utils.cpp
   expect.cpp
+  fake_pruned_blockchain.cpp
   fcmp_pp.cpp
   json_serialization.cpp
   get_xtype_from_string.cpp
